feat: authApi 및 profileApi 파일 추가, 회원가입 및 프로필 조회 기능 구현
